Base controller using mechanism structure
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@1717 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ mechanism/hw_interface/include/hw_interface/hardware_interface.h
@@ -42,7 +42,7 @@ class ActuatorState{
       numEncoderErrors = 0;
     }
   int encoderCount;
-  int timestamp;
+  double timestamp;
   double encoderVelocity;
   bool calibrationReading;
   int lastCalibrationHighTransition;
